International journal of Life science and pharma Research
This is in reference to our journal (International journal of Life science and pharma Research) which is included in your Beals list we have already sent an email last month but we have not received any reply from you. We bring to your kind attention that we have included more peer review process for screening quality articles and to avoid predatory nature. Our journal was included in your list many years back and we have taken this very seriously and have implemented many efforts to reduce the predatory nature. Kindly reevaluate our journal and if found suitable of not predatory please remove from the list. If any process or procedure for getting unlisted from the Beals list kindly let us know.